Removing hooks and validations for muni_id
This seems to have been causing issues… previously, #find_intersecting_municipality was in the after_save hook and a presence validation. Right now, the method is setup to store the information if it doesn’t exist already.

We should update seeds to include muni_ids so that we don’t have a dependency on Carto at run-time. See