Test walkshed generation
In the process of addressing #21 and testing walkshed generation, I discovered a bug, which was caused by a previous commit that updated the projection for a specific column.

This commit cleans up the test suite and adds the first test for a critical feature in the site.

It also adds setup context for the test environment.