Scopes for Models with active Surveys
This commit introduces new model scopes that help with finding records that have active surveys, even when those relations are deeply nested in the association structure.

This will likely introduce some performance issues because of the way these scopes are structured but generally pulling lists of ids from the Survey model (which knows which surveys are active) seems like a more performant solution because the SQL performing inner joins across several tables, then applying a filter. This is an assumption - needs some empirical observation!