Add upper bound to survey set migration range
According to this line here https://github.com/MAPC/myschoolcommute/blob/3b2e4fe43991d97d4cf82f1955f233d959ab43a3/survey/models.py#L320, the SurveySet model originally added a day to the upper bound to increase the scope of the surveys included. Despite the generated reports not doing this for their analyses, the original site adds a day to the range which changes the counts. This change corrects for this so that surveys responses are not missed.