Fix create new school; fix unintended consequences
Setting the default factory changed the way Rails was projecting (or not) INPUT geo data, which began to throw PG errors when PG was getting mismatch SRIDs from Rails. This commit adds an RGeo init which registers a factory for catching other point types that are not part of the default. This may percolate into other problems elsewhere (as the app is untested!).

This commit also goes back in time and changes a migration, which will break any current development versions of databases. It is not clear how to run a migration that changes the coltype using the RGeo activerecord interface.