Add Feature Test for Creating Surveys
This commit adds a feature test for creating surveys. To do this we had to add a distname to the District factory so that the District would show up in the view. We also had to give a user administrative access and add Devise helpers to our test suite. Finally we added the launchy gem to the test group so we can easily debug our feature tests.