Update import task and seed
This commit updates the import script to include the id of the schools from the old database as old_id so that we can join it to the appropriate surveys.

It also updates the import of seeds.rb so that we can bring in the latest surveys and survey responses.