Add rake task to reset sequence id tables in database
This new task was necessary because we kept getting a non-unique record error
when attempting to create new survey responses after deploying to production. It
appears that after ripping down the current production database and then redeploying
when following the instructions outlined in the import.sh file, the sequence ids are
never reset so ActiveRecord attempts to create new records starting at 1 instead of
wherever the old database left off.

If you run into this error, make sure to run command listed in "Step 6" of the import.sh
script.